Move definition of default callbacks to separate header
