+++ src/details/ArborX_DetailsCallbacks.hpp
@@ -1,0 +1,65 @@
+/****************************************************************************
+ * Copyright (c) 2012-2020 by the ArborX authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the ArborX library. ArborX is                       *
+ * distributed under a BSD 3-clause license. For the licensing terms see    *
+ * the LICENSE file in the top-level directory.                             *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+#ifndef ARBORX_DETAILS_CALLBACKS_HPP
+#define ARBORX_DETAILS_CALLBACKS_HPP
+
+#include <Kokkos_Macros.hpp>
+
+namespace ArborX
+{
+namespace Details
+{
+
+struct InlineCallbackTag
+{
+};
+
+struct PostCallbackTag
+{
+};
+
+struct CallbackDefaultSpatialPredicate
+{
+  using tag = InlineCallbackTag;
+  template <typename Query, typename Insert>
+  KOKKOS_FUNCTION void operator()(Query const &, int index,
+                                  Insert const &insert) const
+  {
+    insert(index);
+  }
+};
+
+struct CallbackDefaultNearestPredicate
+{
+  using tag = InlineCallbackTag;
+  template <typename Query, typename Insert>
+  KOKKOS_FUNCTION void operator()(Query const &, int index, float,
+                                  Insert const &insert) const
+  {
+    insert(index);
+  }
+};
+
+struct CallbackDefaultNearestPredicateWithDistance
+{
+  using tag = InlineCallbackTag;
+  template <typename Query, typename Insert>
+  KOKKOS_FUNCTION void operator()(Query const &, int index, float distance,
+                                  Insert const &insert) const
+  {
+    insert({index, distance});
+  }
+};
+
+} // namespace Details
+} // namespace ArborX
+
+#endif
